Board vote to update s2.8 re officer removal
AMEND s2.8 FROM:

An officer of the Board of Directors may be removed from office by the unanimous vote of the remaining officers in a Board of Directors meeting if it the officer in question is found to be in violation of the bylaws, fails to fulfill office responsibilities, or is found in breach of the laws of Maricopa County, the State of Arizona, the United States Federal Government, the Geneva Convention, or the United Federation of Planets.

TO:

An officer of the Board of Directors may be removed from office by the unanimous vote of the remaining officers in a Board of Directors meeting, or by a majority vote of all dues-paid members. A special election must be held within 60 days to elect a replacement according to normal election procedures.